+++ LaborVolumeCalculator/DTO/LaborDto.cs
@@ -5,7 +5,7 @@ namespace LaborVolumeCalculator.DTO
         public int ID { get; set; }
         public string Code { get; set; }
         public string Name { get; set; }
-        public float MinVolume { get; set; }
-        public float MaxVolume { get; set; }
+        public double MinVolume { get; set; }
+        public double MaxVolume { get; set; }
     }
 }
